Fix incorrect version of WebSocketClient instance being used.
API v4 requires the use of NewWebSocketClient4 instead
NewWebSocketClient.
Now that APIv3 is still supported this is not an issue, but this hides
the problem until the APIv3 is officially dropped. Where trying to
connect to APIv4 using this constructor will lead to a panic without
giving clear information about the problem ("websocket: bad handshake").